requirements/requirements.txt
email-validator==1.3.0
Flask==2.2.2
requests==2.28.1
slack-sdk==3.18.3
python-dotenv==0.21.0

